refactor: StoreInfo -> store_info_model 이름 변경
