+++ lib/models/store_info_model.dart
